Refactors entire Filter component into separate component file, makes the buttons fully responsive with layout, just need to pass props in for desktop and mobile.
